Add files via upload
This project aims to build a robust fraud detection system capable of accurately identifying fraudulent transactions while minimizing false positives to prevent inconvenience to legitimate customers. The exploration of this dataset serves as a foundation for developing and deploying effective fraud detection models. #The dataset contains numerical input variables that are the result of a PCA transformation due to confidentiality issues. Features V1, V2, ..., V28 are the principal components obtained from the PCA transformation, and 'Time' and 'Amount' are the only features that have not been transformed. #The 'Time' feature represents the time elapsed between each transaction and the first transaction in the dataset. The 'Amount' feature is the transaction amount. The 'Class' feature takes value 1 in case of fraud and 0 otherwise. #The project based on this dataset involves exploring various aspects such as:

#Data Preprocessing: Handling missing values, scaling features, dealing with imbalanced classes. #Exploratory Data Analysis (EDA): Understanding the distribution of transactions, exploring the relationship between features, investigating the class imbalance, and identifying potential patterns in fraudulent transactions. #Model Building: Developing predictive models (e.g., Logistic Regression, Random Forest, Gradient Boosting, Neural Networks) to classify transactions as fraudulent or non-fraudulent. #Model Evaluation: Assessing model performance using metrics like accuracy, precision, recall, F1-score, and area under the ROC curve (AUC-ROC) considering the class imbalance. #Handling Imbalanced Classes: Applying techniques like oversampling (SMOTE), undersampling, or using class weights to address the imbalance issue. #Hyperparameter Tuning: Optimizing model parameters to improve performance. #Deployment and Monitoring: Implementing the chosen model into a production environment and continuously monitoring its performance.